+++ templates/snap/run-snap-graph.sh
@@ -32,10 +32,15 @@ source **caroline_virtual_environment_directory**/bin/activate
 echo "$(date '+%Y-%m-%dT%H:%M:%S'): $(whoami) in $(pwd) has started run-snap-graph.sh (ARRAY ID $SLURM_ARRAY_TASK_ID) (AoI **snap:general:AoI-name**, track **track**) with slurm-ID $SLURM_JOB_ID)" >> **caroline_work_directory**/submitted_jobs.log
 
 output_path="**snap-output-path**"
+ROME_CONSTRAINED="**rome-constrained**"
 
 graph=`ls ${output_path}/PROCESSID-${SLURM_ARRAY_TASK_ID}-*.xml`
 
-gpt_query="gpt -q ${SLURM_CPUS_PER_TASK} -c $((9 * $SLURM_CPUS_PER_TASK))G -J-Xmx$((15 * $SLURM_CPUS_PER_TASK))G -e -x -J-Dsnap.jai.defaultTileSize=512 -J-Dsnap.dataio.reader.tileWidth=512 -J-Dsnap.dataio.reader.tileHeight=512"
+if [ "${ROME_CONSTRAINED}" -eq "1" ]; then  # we have 16GB cores
+  gpt_query="gpt -q ${SLURM_CPUS_PER_TASK} -c $((9 * $SLURM_CPUS_PER_TASK))G -J-Xmx$((15 * $SLURM_CPUS_PER_TASK))G -e -x -J-Dsnap.jai.defaultTileSize=512 -J-Dsnap.dataio.reader.tileWidth=512 -J-Dsnap.dataio.reader.tileHeight=512"
+else # we have 12GB cores
+  gpt_query="gpt -q ${SLURM_CPUS_PER_TASK} -c $((7 * $SLURM_CPUS_PER_TASK))G -J-Xmx$((11 * $SLURM_CPUS_PER_TASK))G -e -x -J-Dsnap.jai.defaultTileSize=512 -J-Dsnap.dataio.reader.tileWidth=512 -J-Dsnap.dataio.reader.tileHeight=512"
+fi
 
 ${gpt_query} --diag